refactor: break up touchbackend code
